requirements.txt
torch>=2.3.0
transformers>=4.51.0
trl>=0.19.0
peft>=0.12.0
accelerate>=0.34.0
datasets>=2.21.0
pandas>=2.2.0
pyyaml>=6.0.1
tqdm>=4.66.0
